Bug fixes * Displaying of prices is now via 'floatval()' instead of 'intval()' * User sign-up javascript bla bla * In Seat_model and Slot_model particularly, many functions have added parameters so as to narrow down the search. * I've lost track. HUHUHU
New features
* CHANGE SHOWING TIME!!!!
* UPGRADE TICKET CLASS!!!!
* VIEW DETAILS OF BOOKING WHEN CHANGES ARE PENDING (I.E. payment pending ).
* CANCEL CHANGES TO BOOKING  (THUS REVERT TO ORIGINAL STATE) WHEN CHANGES ARE PENDING (I.E. payment pending ).
* Introduction of transactionList_model - records transactions of users
* Added new "Metro tiles", for manage booking still

Changes in Features/Algorithms
* Passwords are now in whirlpool-hash!!! Finally!
* User sign-up validation now resembles book step 3 guest form validation
* Deprecated user of book_step session data. Uses now combination 
of 'activity_name' and 'activity_stage'
* Deprecating the use of CoordinateSecurity_model and all associated operations, thus
* All cookies (except for 'x_slot_UUIDs' ) and session data are set, got, and deleted via respective calls to clientsidedata_model .

+++ trunk/application/views/confirmReservation/confirmReservation02-free.php
@@ -168,7 +168,7 @@ $this->load->view('html-generic/metadata.inc');
 											<td><?php echo $singlePurchase->Charge_type_Description; ?></td>
 											<td>
 											<?php
-												$thisItemAmount = intval($singlePurchase->Amount); 
+												$thisItemAmount = floatval($singlePurchase->Amount); 
 												$totalCharges += $thisItemAmount;
 												if ( $thisItemAmount < 0 )
 													echo '('.$thisItemAmount.')';
